Made helper class and tests.
